% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stpmixtdata.R
\name{generate_sptmixtdata}
\alias{generate_sptmixtdata}
\title{Create spatio temporal data}
\usage{
generate_sptmixtdata(
  min_date = as.Date("01-01-2009", format = "\%d-\%m-\%Y"),
  max_date = as.Date("31-12-2010", format = "\%d-\%m-\%Y"),
  frequency = "month",
  nSite = 8,
  nLandUse = 2,
  nBasis = 2,
  betaTrue = NULL,
  parameters = NULL,
  missingMeasures = list(status = F, ratio = 0, type = "MNAR")
)
}
\arguments{
\item{min_date}{date minimum data for the simulated data}

\item{max_date}{date maximum data for the simulated data}

\item{frequency}{string Observation frequency. Default to "month".}

\item{nSite}{numeric Number of locations.}

\item{nLandUse}{numeric Number of different landuses.}

\item{nBasis}{numeric Number of temporal basis functions.}

\item{betaTrue}{numeric matrix Range of variation. Default to -2 to 2.}

\item{parameters}{list Uncertainty parameters. Default to sigma (1), tau (0.5), phi (1).}

\item{missingMeasures}{list Frequency and type of missing data. Default to no missing measures.}
}
\description{
Create fake data following a cluster of spatio-temporal signals
}
