% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tempbasis.R
\name{generate_tempbasis}
\alias{generate_tempbasis}
\title{Defining the temporal basis}
\usage{
generate_tempbasis(
  data,
  tempBasis_fun = "re",
  tempPeriod = "\%m",
  nSplines = 3,
  splinesType = "poly"
)
}
\arguments{
\item{data}{data.frame}

\item{tempBasis_fun}{string One of "re" (random effect) or "bs" (splines)}

\item{tempPeriod}{string Time unit (default to "\%m" - month)}

\item{nSplines}{numeric Number of splines used to define the temporal basis functions}

\item{splinesType}{string One of "poly", "penalized", "cubic"}
}
\description{
Use the data range and creates the basis functions of the temporal model
}
