% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptmixtR-package.R
\docType{package}
\name{sptmixtR-package}
\alias{sptmixtR}
\alias{sptmixtR-package}
\title{sptmixtR: Mixture models with spatio-temporal dependencies}
\description{
This package provide the necessary functions to model a mixture distribution with spatio-temporal observations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ick003/sptmixtR}
  \item Report bugs at \url{https://github.com/ick003/sptmixtR/issues}
}

}
\author{
\strong{Maintainer}: Adrien Ickowicz \email{adrien.ickowicz@data61.csiro.au} (\href{https://orcid.org/0000-0001-7861-7340}{ORCID})

}
\keyword{internal}
